Refactor RoiClickAndDragMode into separate module
